Added support for retrieving certificates when asynchronous order finalization is enabled on the ACME server-side. // see: https://community.letsencrypt.org/t/enabling-asynchronous-order-finalization/193522
Description: The `getCertificate()` method in the `Client` class has been updated to handle cases where the certificate is not immediately available and retries fetching the certificate until it becomes valid.
This commit also introduces a new (private) method `getCertificateChain()` in the `Client` class, which is used to fetch the certificate chain from the ACME API to avoid implementing the same code multiple times.
Additionally, the `Order` class has been updated to include the new `certificate` property and a corresponding `getCertificate()` getter method.